Class TestBase is created
Common methods are moved to the class TastBase
