+++ addressbook-web-tests/src/test/java/pk/addressbook/GroupCreationTests.java
@@ -1,43 +1,7 @@
 package pk.addressbook;// Generated by Selenium IDE
 import org.junit.Test;
-import org.junit.Before;
-import org.junit.After;
+public class GroupCreationTests extends TestBase {
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.core.IsNot.not;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.JavascriptExecutor;
-
-import java.util.*;
-
-public class GroupCreationTests {
-  private WebDriver driver;
-  private Map<String, Object> vars;
-  JavascriptExecutor js;
-  @Before
-  public void setUp() {
-    driver = new FirefoxDriver();
-    js = (JavascriptExecutor) driver;
-    vars = new HashMap<String, Object>();
-    driver.get("http://localhost/addressbook/");
-    /*driver.manage().window().setSize(new Dimension(550, 691));*/
-    login("admin", "secret");
-  }
-
-  private void login(String username, String password) {
-    driver.findElement(By.name("user")).sendKeys(username);
-    driver.findElement(By.id("LoginForm")).click();
-    driver.findElement(By.name("pass")).click();
-    driver.findElement(By.name("pass")).sendKeys(password);
-    driver.findElement(By.cssSelector("input:nth-child(7)")).click();
-  }
-
-  @After
-  public void tearDown() {
-    driver.quit();
-  }
   @Test
   public void testGroupCreation() {
     goToGroupPage();
@@ -47,24 +11,4 @@ public class GroupCreationTests {
     goToGroupPage();
   }
 
-  private void submitGroupCreation() {
-    driver.findElement(By.name("submit")).click();
-  }
-
-  private void fillGroupForm(GroupData groupData) {
-    driver.findElement(By.name("group_name")).click();
-    driver.findElement(By.name("group_name")).sendKeys(groupData.name());
-    driver.findElement(By.name("group_header")).click();
-    driver.findElement(By.name("group_header")).sendKeys(groupData.header());
-    driver.findElement(By.name("group_footer")).click();
-    driver.findElement(By.name("group_footer")).sendKeys(groupData.footer());
-  }
-
-  private void initGroupCreation() {
-    driver.findElement(By.name("new")).click();
-  }
-
-  private void goToGroupPage() {
-    driver.findElement(By.linkText("groups")).click();
-  }
 }
